backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.1.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
clr-loader==0.2.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dlib==19.24.4
et-xmlfile==1.1.0
executing==2.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
faiss-cpu==1.9.0
fastjsonschema==2.19.0
filelock==3.14.0
fiona==1.9.6
flack==1.3.7
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.45.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.6.0
geopandas==0.14.4
google-auth==2.31.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.64.1
gTTS==2.5.1
h11==0.14.0
h5py==3.11.0
hdf5plugin==4.4.0
htmlmin==0.1.12
httpcore==1.0.6
httpx==0.27.2
hydra-core==1.3.2
idna==3.6
ImageHash==4.3.1
imageio==2.32.0
imageio-ffmpeg==0.4.9
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
japanize-matplotlib==1.1.3
jedi==0.19.1
Jinja2==3.1.2
jiter==0.6.1
joblib==1.4.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langsmith==0.1.136
lazy_loader==0.3
libclang==18.1.1
llvmlite==0.42.0
macholib==1.16.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mecab-python3==1.0.9
mistune==3.0.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
multimethod==1.11.2
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.59.1
numpy==1.26.1
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.52.0
opencv-python==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.7
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.2.0
platformdirs==4.2.2
plotly==5.22.0
portalocker==2.8.2
proglog==0.1.10
prometheus-client==0.19.0
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydotplus==2.0.2
pygame==2.5.2
Pygments==2.18.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pyparsing==3.1.1
pypdf==5.0.1
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pythonnet==3.0.3
pytz==2023.3.post1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
regex==2024.9.11
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.13.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.11.3
seaborn==0.13.2
Send2Trash==1.8.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spicy==0.16.0
SQLAlchemy==2.0.36
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
tenacity==8.3.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.15.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.5.0
tifffile==2023.9.26
tiktoken==0.8.0
tinycss2==1.2.1
torch==2.3.0
torchtext==0.18.0
torchvision==0.18.0
tornado==6.4.1
touch==2020.12.3
tqdm==4.66.1
traitlets==5.14.3
typeguard==4.2.1
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
values==2020.12.3
visions==0.7.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.3
widgetsnbextension==4.0.9
wordcloud==1.9.3
wrapt==1.14.1
xgboost==2.0.3
yarl==1.15.5
ydata-profiling @ https://github.com/ydataai/pandas-profiling/archive/master.zip#sha256=2b74e0759fc50b6707bd848752d96205d800ee68bbf82f201a47f7e8cb66fd91
